changed graph_mpi_weight.hpp a bit. Will make drastic changes tomorrow

+++ max-flow/serial_correct/max_flow.cpp
@@ -1,7 +1,7 @@
 #include "max_flow_csr.hpp"
 #include <fstream>
 
-int32_t  main (long long  argc, char ** argv) {
+int32_t  main (int argc, char ** argv) {
 
 
   char* file_name = argv[1];
